# Conflicts:
#	src/main/java/player/Human.java
#	src/main/java/player/Ogre.java

+++ src/main/java/util/Killable.java
@@ -3,7 +3,7 @@ package util;
 import player.Player;
 
 public interface Killable {
-    void takeDmg(Player by, int amount);
+    void takeDmg(Player by, double amount);
 
     void die();
     int getXP();
